Restructured the tree of the project(added one more package). Added @Builder and @Service annotations to class ServeAdServiceImpl
