+++ src/main/java/com/mabaya/exam/SpringBootExamApplication.java
@@ -7,6 +7,7 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 public class SpringBootExamApplication {
 
     public static void main(String[] args) {
+
         SpringApplication.run(SpringBootExamApplication.class, args);
     }
