+++ src/main/java/com/mabaya/exam/app/dao/CampaignRepository.java
@@ -1,6 +1,6 @@
-package com.mabaya.exam.dao;
+package com.mabaya.exam.app.dao;
 
-import com.mabaya.exam.model.Campaign;
+import com.mabaya.exam.app.model.Campaign;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
 public interface CampaignRepository extends MongoRepository<Campaign, String> {
