+++ src/main/java/com/mabaya/exam/app/dto/AddCampaignDto.java
@@ -1,4 +1,4 @@
-package com.mabaya.exam.dto;
+package com.mabaya.exam.app.dto;
 
 import lombok.*;
