--- src/main/java/com/mabaya/exam/service/CampaignService.java
@@ -1,9 +1,0 @@
-package com.mabaya.exam.service;
-
-import com.mabaya.exam.dto.AddCampaignDto;
-import com.mabaya.exam.dto.CreatedCampaignResponseDto;
-
-public interface CampaignService {
-
-    CreatedCampaignResponseDto addNewCampaign(AddCampaignDto newCampaign);
-}
