--- src/main/java/com/mabaya/exam/service/ServeAdServiceImpl.java
@@ -1,10 +1,0 @@
-package com.mabaya.exam.service;
-
-import com.mabaya.exam.dto.ServeAdDto;
-
-public class ServeAdServiceImpl implements ServeAdService {
-    @Override
-    public ServeAdDto promotedProduct(String category) {
-        return null;
-    }
-}
